Added some stuff
Added colour coding ro the character blueprint

Added health bar functionality (lose health by pressing F)

Added a function to the gamemode so that if the player is destroyed (0 health), the level restarts.

Added a spawner trigger so moths start to spawn once you move past a certain point.

Removed looping feature on spawner, will be readded with editable paramaters like number of waves to spawn and waiting for all of one wave to be destroyed before spawning another